Added more s3 mocking to fix broken testcases
